run/requirements.old.txt
aniso8601==9.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cssmin==0.2.0
cycler==0.10.0
firebase-admin
flake8==3.9.2
Flask==2.0.1
Flask-Assets==2.0
Flask-Caching==1.10.1
Flask-RESTful==0.3.9
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-bigquery==2.24.1
google-cloud-core==2.0.0
google-cloud-firestore
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsmin==2.2.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
numpy==1.21.2
packaging==21.0
pandas==1.3.2
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
timezonefinder==5.2.0
torch==1.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
utm==0.7.0
webassets==2.0
Werkzeug==2.0.1
